Changed AppComponent to load from an ITrack object.
Added an ITrack object, which encapsulates a track to play.
Added app.template.html, to provide the View for the AppComponent.
Modified AppComponent to use a separate file for the View, and to store an ITrack object of the current track being played.
Added nack-deep.jpg as a sample album art cover.

Modified index.html to have a loading spinner.

Added app.scss to provide styles for the AppComponent.
Added mixins.scss to provide common mixins.
Added fonts.scss to provide common fonts.
Modified main.scss to include all of the new items.
Added Helvetica and Helvetica Bold.